requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.1
cachetools==5.3.2
celery==5.4.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==2023.10
cssselect==1.2.0
cssutils==2.9.0
Django==5.0
django-axes==6.2.0
django-crispy-forms==2.1
django-redis==5.4.0
dotenv==0.0.5
gunicorn==21.2.0
idna==3.6
kombu==5.4.2
lxml==4.9.4
packaging==23.2
premailer==3.10.0
prompt_toolkit==3.0.48
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.1.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
yagmail==0.15.293
